+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2015-12-08
 - update `boto` to 1.2.2
+- fix for wilcard default name filter in `ebzl instances`
 
 ## 2015-12-05
 - use a wildcard as a default name filter in `ebzl instances` (`--name="*"`)